requirements.txt
# 数据库
pymysql==1.1.0

# 配置管理
python-dotenv==1.0.0

# LangChain 核心依赖
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# 向量数据库
faiss-cpu==1.8.0

# FastAPI 服务
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# 日志 (可选，用于 JSON 格式日志)
python-json-logger==2.0.7
